Meta/Core/Various: Switch to C++20 as a requirement, Require Qt6, remove all qt5 related stuff, remove tcb::span use std::span, deprecation fixes. etc.

+++ include/inviwo/qt/editor/linkdialog/linkdialogprocessorgraphicsitems.h
@@ -54,7 +54,9 @@ public:
     virtual int getLevel() const override;
 
     // override for qgraphicsitem_cast (refer qt documentation)
-    enum { Type = UserType + LinkDialogProcessorGraphicsItemType };
+    enum {
+        Type = static_cast<int>(UserType) + static_cast<int>(LinkDialogProcessorGraphicsItemType)
+    };
     virtual int type() const override { return Type; }
     virtual void updatePositions() override;
